Comments
My name is Shauna, so the first and last letter of my name (S + A = 19 + 1) gives the number 20. The 20th Fibonacci number is 6,765.

My surname is Byrne
The first letter B is number 66
The last letter e is number 101
Fibonacci number 167 is 35600075545958458963222876581316753
The ord() generally indicates a number defining the position of a value in a series, for example the 3rd value in the series.