Create Iris.py
Added the headings & dash to give the information a cleaner format. Also this allows any user looking at the data to know which information is in each column as they have been re-arranged to fit the listing in the assignment. Program gives required numerical data without printing the headers also.